Added little face
Added a little face, getting the drawCircle to work. For some reason, the drawCircle function doesn't work well in the EVERY_N_MILLISECONDS function, and has to be updated outside on it's own.
The screen is very glitchy, oftentimes the white eyes shudder and the initial loading screen gets stuck in the background.

+++ src/FourDuino.cpp
@@ -790,7 +790,7 @@ bool OLED::drawCircle(uint16_t x, uint16_t y, uint16_t radius, uint16_t color)
     return getAck();
 }
 
-bool OLED::drawCircle(uint16_t x, uint16_t y, uint16_t radius, Color color)
+bool OLED::drawCircle(uint8_t x, uint8_t y, uint8_t radius, Color color)
 {
     return drawCircle(x, y, radius, color.to16BitRGB());
 }
